requirements.txt
flask                
numpy                
pandas               
torch                
open-clip-torch      
scikit-learn         
pillow               